+++ maser/data/pds/ppi/__init__.py
@@ -1,0 +1,1 @@
+from maser.data.pds.ppi import *
