+++ maser/data/pds/ppi/voyager/__init__.py
@@ -1,0 +1,1 @@
+from maser.data.pds.ppi.voyager import *
