--- maser/data/pds/voyager/const.py
@@ -1,19 +1,0 @@
-from .pra import PDSPPIVoyagerPRAHighRateDataTimeSeriesDataFromLabel, PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel
-
-__all__ = ['PDS_OBJECT_CLASSES']
-
-PDS_OBJECT_CLASSES = {
-    'VG1-J-PRA-3-RDR-LOWBAND-6SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG1-S-PRA-3-RDR-LOWBAND-6SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-J-PRA-3-RDR-LOWBAND-6SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-S-PRA-3-RDR-LOWBAND-6SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-N-PRA-3-RDR-LOWBAND-6SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-U-PRA-3-RDR-LOWBAND-6SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG1-J-PRA-4-SUMM-BROWSE-48SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG1-S-PRA-4-SUMM-BROWSE-48SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-J-PRA-4-SUMM-BROWSE-48SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-S-PRA-4-SUMM-BROWSE-48SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-N-PRA-4-SUMM-BROWSE-48SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-U-PRA-4-SUMM-BROWSE-48SEC-V1.0': PDSPPIVoyagerPRARDRLowBand6SecDataFromLabel,
-    'VG2-N-PRA-2-RDR-HIGHRATE-60MS-V1.0': PDSPPIVoyagerPRAHighRateDataTimeSeriesDataFromLabel,
-}
